requirements.txt
acme==1.12.0
attrs==20.3.0
blinker==1.6.2
certbot==1.12.0
certbot-apache==1.10.1
certifi==2020.6.20
chardet==4.0.0
click==8.1.3
ConfigArgParse==1.2.3
configobj==5.0.6
cryptography==41.0.0
distro==1.5.0
distro-info==1.0
Flask==2.3.2
Flask-Cors==3.0.10
httplib2==0.18.1
idna==2.10
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
josepy==1.2.0
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
MarkupSafe==2.1.3
more-itertools==4.2.0
mysql-connector-python==8.0.33
oauthlib==3.1.0
parsedatetime==2.6
protobuf==3.20.3
pycurl==7.43.0.6
PyJWT==2.4.0
pyOpenSSL==20.0.1
pyRFC3339==1.1
pyrsistent==0.15.5
PySimpleSOAP==1.16.2
python-augeas==0.5.0
python-debian==0.1.39
python-debianbts==3.1.0
pytz==2021.1
PyYAML==5.4
requests==2.25.1
requests-toolbelt==0.9.1
six==1.16.0
urllib3==1.26.5
Werkzeug==2.3.6
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.2.0
schedule==1.2.0
